requirements.txt
flask
tensorflow
numpy
pillow
scikit-learn
mysql-connector-python
gunicorn